refactor(social-inbox): dedupe interfaces, executor helper, node union
- Move ~15 social-inbox domain interfaces/type aliases out of
  packages/services/social/messages.service.ts into
  packages/interfaces/src/social/social-inbox.interface.ts, and the two
  model classes into packages/models/social/, matching the sibling
  link.interface/link.model pattern. Update messages.service.ts and the
  messages-page.tsx consumer imports; barrel the new interface file.
- Hoist the verbatim-duplicated getConfigOrInputString helper from the
  post-reply and send-dm SaaS executors into BaseExecutor as a protected
  method; both already extend it.
- Remove the duplicate 'commentTrigger' member from the SaaSNodeType union
  (auto-merge leftover), keeping it in its correct trigger group.

Behavior-preserving. Focused type-checks pass across interfaces, models,
services, workflow-engine, workflow-saas, and app; executor + registry
specs green.
