+++ packages/interfaces/src/social/index.ts
@@ -1,1 +1,2 @@
 export * from './link.interface';
+export * from './social-inbox.interface';
